==> 08-best_worst_scaling.qmd <==
# Best-Worst Scaling

```{r include=FALSE}
library(tidyverse)
library(crossdes)
library(support.BWS)
library(dfidx)
library(mlogit)
```


Best–worst scaling (BWS) is a method for measuring preferences among pre-defined items.^[Notes are from http://lab.agr.hokudai.ac.jp/nmvr/03-bws1.html.] BWS presents subsets of the items for evaluation, asking the respondent to identify the best and worst (or most important and least important) from the subset. BWS studies usually use a balanced incomplete block design (BIBD) to construct choice sets.

BIBD is a category of designs where `trt` treatments (items) are presented to the survey respondent in `b` blocks (survey questions) of `k` treatments (items). For example, seven items might be presented in seven survey questions with four of the items show in each question.

```{r}
items <- c("origin", "variety", "price", "taste", "safety", "washfree", "milling")

set.seed(8041)
bibd <- find.BIB(
  trt = length(items),
  b = length(items),
  k = 4
)
```

The documentation for `isGYD()` explains design balance. A balanced block design has three qualities: i) Each treatment appears equally often. ii) Each treatment appears in each block either *n* or *n+1* times (usually 0 or 1 times). iii) The concurrences of *i* and *j* are the same for all pairs of treatments (*i*, *j*). A design balanced with respect to both rows and columns is called a generalized Youden design (GYD). A design with less columns (rows) than treatments is *incomplete* with respect to rows (columns). That's what you will have in a BWS. A design in which each treatment occurs once per row and column is a *latin square*. If each treatment occurs the same number of times per row and column, it is a *generalized latin square*. 

```{r}
isGYD(bibd)
```

There are other design with a BIBD: 

```{r eval=FALSE}
find.BIB(6, 10, 3) %>% isGYD()   # 6 items shown in 10 blocks of 3
find.BIB(7, 7, 3) %>% isGYD()    # 7 items shown in 7 blocks of 3
find.BIB(9, 12, 6) %>% isGYD()   # 9 items shown in 12 blocks of 6
find.BIB(11, 11, 5) %>% isGYD()  # 11 items shown in 11 blocks of 5
find.BIB(13, 13, 4) %>% isGYD()  # 13 items shown in 13 blocks of 4
find.BIB(16, 16, 6) %>% isGYD()  # 16 items shown in 16 blocks of 6
```

The resulting questionaire might look like this.

```{r}
bws.questionnaire(bibd, design.type = 2, item.names = items)
```

Consider the following response data from a BIBD with 7 items shown in 7 blocks of 3. `b1` is the item selected as best in question 1, `w1` is the item selected as worst in question 1, etc. `age`, `hp`, and `chem` are respondent covariates.

```{r}
data("ricebws1", package = "support.BWS")
dat <- ricebws1
glimpse(dat)
```

Convert each response into one row per possible best-worse pair. There are `k(k-1)` possible pairs, in this case 4x3=12 pairs. `bws.dataset()` lengthens the 90 rows to 90 x 7 questions x 12 possible pairs per question = 7,560 rows.

```{r}
bws <- bws.dataset(
  data = dat,
  response.type = 1, # format of response variables: 1 = row number format
  choice.sets = bibd,
  design.type = 2, # BIBD
  item.names = items,
  id = "id", # respondent id variable
  response = colnames(dat)[2:15], # response variables
  model = "maxdiff" #  type of dataset to create
) 

# 90 respondents x 7 questions x 12 possible pairs per question
dim(bws)
```

Question 1 presented items [`r paste(items[bibd[1, ]], collapse = ", ")`]. Respondent `id` = 1 selected choice 4 (`items[bibd[1, 4]]` = `r items[bibd[1, 4]]`) for the Best (`b1`) and choice 1 (`items[bibd[1, 1]]` = `r items[bibd[1, 1]]`) for Worst (`w1`). 

```{r}
dat[1, ]
```

The converted data set is easier to translate. Column `RES = TRUE` in row 10 indicates the pair the respondent selected. Best was item 7 (`items[7]` = `r items[7]`) and the Worst was item 2 (`items[2]` = `r items[2]`). You can also get that from the +1 and -1 indicators for `milling` and `variety` used for modeling.

```{r}
bws %>% filter(id == 1, Q == 1)
```

## Count Analysis

You can see below that respondent 1 selected item 1 (`r items[1]`) Best three times, item 5 (`r items[5]`) Best three times, and item 7 (`r items[7]`) Best once. Respondent 1 selected items 2 (`r items[2]`), 3 (`r items[3]`), 4 (`r items[4]`), and 5 (`r items[5]`) Worst once and item 6 (`r items[6]`) worst three times.

```{r}
bws %>% 
  filter(id == 1, RES == TRUE) %>%
  gtsummary::tbl_cross(BEST, WORST)
```

`bws.count()` calculates counts for (b)est, (w)orst, best-minus-worst (bw), and standardized bw (sbw = bw / number of levels) for each item.

```{r}
bws_count <- bws.count(bws, cl = 2)
dim(bws_count)
```



```{r}
bws_count %>% filter(id == 1) %>% glimpse()
```

`plot()` shows the relationship between the level means and standard deviations. Price, taste, and safety are similarly important, but price has a higher standard deviation, meaning its importance varies.

```{r}
plot(bws_count, score = "bw")
```

The column plot shows the item ranks.

```{r}
bws_count %>%
  select(id, starts_with("sbw")) %>%
  pivot_longer(cols = -id) %>%
  group_by(name) %>%
  summarize(.groups = "drop", M = mean(value)) %>%
  arrange(M) %>%
  ggplot(aes(y = fct_inorder(name), x = M)) +
  geom_col()
```

## Model

Fit a conditional logit model. A simple model uses the dummy vars, excluding one (washfree) to avoid singularity. The last term “- 1” means that the model has no alternative-specific constants. Use `dfidx()` to convert the data into a format appropriate for the model.

```{r}
fmla <- RES ~ origin + variety + price + taste + safety + milling - 1

bws_dfidx <- dfidx(bws, idx = list(c("STR", "id"), "PAIR"), choice = "RES")

mlogit_fit <- mlogit(formula = fmla, data = bws_dfidx)

summary(mlogit_fit)
```

`bws.sp()` shows the shares of preference.

```{r}
# Specify the name of the base since it isn't in model.
(bws_sp <- bws.sp(mlogit_fit, base = "washfree"))
```

Safety was most important and was `0.252 / 0.238` = `r bws_sp["safety"] / bws_sp["price"]` times as important as the second place price.

This model isn't a great fit, unfortunately. You cannot pull the McFadden's R-squared easily, but the calculation is straight-forward.

```{r collapse=TRUE}
ll0 <- -90 * 7 * log(12) # log-likelihood at zero
llb <- as.numeric(mlogit_fit$logLik)
1 - (llb/ll0) # McFadden's R-squared
1 - ((llb-6)/ll0) # Adjusted McFadden's R-squared
```

A possible improvement is the random parameters logit model. 

```{r}
fmla_rp <- RES ~ origin + variety + price + taste + safety + milling - 1 | 0

mlogit_rp_fit <- mlogit(
  fmla_rp,
  bws_dfidx,
  rpar = c(origin = "n", variety = "n", price = "n", taste = "n", safety = "n", milling = "n"),
  R = 100,
  halton = NA,
  panel = TRUE
)

summary(mlogit_rp_fit)
```

McFadden's R-squared increased substantially.

```{r collapse=TRUE}
llb_rp <- as.numeric(mlogit_rp_fit$logLik)
1 - (llb_rp/ll0) # McFadden's R-squared
1 - ((llb_rp-6)/ll0) # Adjusted McFadden's R-squared
```

```{r}
# Specify the name of the base since it isn't in model.
(bws_rp_sp <- bws.sp(mlogit_rp_fit, base = "washfree", coef = items[-6]))
```

Now (surprisingly?), price is most important and was `0.411 / 0.286` = `r bws_rp_sp["price"] / bws_rp_sp["safety"]` times as important as the second place safety.
