==> 09-case_study_nhanes.qmd <==
# Case Study: NHANES

This case study follows the tutorial laid out in section 7.5 of @Heeringa2017. It uses the *National Health and Nutrition Examination Survey (NHANES)* to build a predictive model of diastolic blood pressure (DBP). The NHANES sample is a stratified cluster sample, weighted for unequal probabilities of selection, non-response, and post-stratification.

```{r setup, message=FALSE, warning=FALSE}
library(tidyverse)
library(scales)
library(janitor)
library(glue)
library(survey)
library(srvyr)
library(gtsummary)
library(broom)
library(ggeffects)
library(svydiags)
library(NHANES)
```

## Data Preparation

```{r}
data(NHANESraw)
```

Heeringa identifies a data cleaning step, converting 0s to `NA`.

```{r}
nhanes <-
  NHANESraw |> 
  filter(SurveyYr == "2011_12") |> 
  mutate(
    BPDia1 = if_else(BPDia1 == 0, NA_integer_, BPDia1),
    Race1 = fct_relevel(Race1, "Mexican", after = 0),
    Gender = fct_relevel(Gender, "male", after = 0)
  )
```

Create a survey design object for analysis. Use it to create a new variable, centered age, `Age`.

```{r}
nhanes_des <- as_survey_design(
  nhanes,
  ids = SDMVPSU,  # primary sampling unit (cluster)
  strata = SDMVSTRA,  # sampling stratum
  nest = TRUE,  # strata ids are repeated across clusters
  weights = WTMEC2YR
)

weighted_mean_age <-
  nhanes_des |>
  filter(Age >= 18) |>
  summarize(M = survey_mean(Age)) |>
  pull(M)

nhanes_des <- 
  nhanes_des |> 
  mutate(
    Age18p = Age >= 18,
    Age = Age - weighted_mean_age
  )

summary(nhanes_des)
```

## Descriptive Analysis

Regress the continuous response variable, diastolic blood pressure (`BPDia1`), on four predictors of interest: race/ethnicity (`Race1`), centered-age (`Age`), gender (`Gender`), and sexual orientation (`SexOrientation`).

```{r}
#| code-fold: true

nhanes_des |>
  tbl_uvregression(
    y = BPDia1,
    method = survey::svyglm,
    method.args = list(subset = (Age18p == 1)),
    include = c(Race1, Age, Gender, SexOrientation),
    estimate_fun = label_style_number(digits = 2),
    pvalue_fun = label_style_pvalue(digits = 2),
    label = list(
      Race1 ~ "Race/ethnicity", 
      Age ~ "Age (cent.)", 
      Gender ~ "Gender"
      # MaritalStatus ~ "Marital status"
    )
  ) |>
  # add_global_p(anova_fun = tidy_wald_test) |>
  as_gt() |>
  gt::tab_caption(".") |>
  gt::tab_header(
    title = "Initial Design-Based Bivariate Regression Analysis Results ",
    subtitle = glue("Assessing Potential Predictors of Diastolic Blood Pressure ",
                    "for the 2011–2012 NHANES Adult Sample")
  ) |>
  gt::tab_options(heading.align = "left") 
```

The design-based t-tests suggest that race/ethnicity, age, and gender have  potentially significant relationships with DBP, while sexual orientation does not appear to be related. Non-Hispanic whites, non-Hispanic blacks, older adults, and males appear to have the highest DBPs at first glance. Include the first three predictors in an initial model. 

## Analysis

### Model Fitting

Fit an initial regression model with the variables of interest. Notice how the dataset is filtered _inside_ the model with `subset`.

```{r}
fit_1 <- svyglm(
  BPDia1 ~ Race1 + Age + Gender,
  design = nhanes_des,
  subset = (Age18p == 1)
)
```

```{r}
#| code-fold: true
fit_1 |>
  gtsummary::tbl_regression(
    label = list(
      Race1 ~ "Race/ethnicity", 
      Age ~ "Age (cent.)", 
      Gender ~ "Gender"
    ),
    intercept = TRUE,
    estimate_fun = label_style_number(digits = 3)
  ) |>
  gtsummary::add_glance_source_note() |>
  gtsummary::add_n(location = "level") |>
  gtsummary::bold_p() |>
  as_gt() |>
  gt::tab_caption(".") |>
  gt::tab_header(
    title = "Initial Model for Diastolic Blood Pressure",
    subtitle = glue("Design-Adjusted Test Statistics and Confidence Intervals ",
                    "for the Parameters, and Design Effects for the Parameter ",
                    "Estimates.")
  ) |>
  gt::tab_options(heading.align = "left")
```

The *t* statistics degrees of freedom equal the number of ultimate clusters (31) minus the number of strata (14).

Plot residuals against the predictor variables. The `augment()` function has not been specifically written for `svyglm` objects, so you need to make a couple adjustments. Start with the centered age, `Age`. The plot has a curvilinear pattern suggesting a quadratic relationship. Add a squared term and re-plot the residuals.

```{r}
preds_1 <-
  augment(fit_1) |>
  mutate(
    .se.fit = sqrt(attr(.fitted, "var")),
    .fitted = as.numeric(.fitted)
  )

fit_2 <- svyglm(
  BPDia1 ~ Race1 + Age + I(Age^2) + Gender,
  design = nhanes_des,
  subset = (Age18p == 1)
)

preds_2 <- 
  augment(fit_2) |>
  mutate(
    .se.fit = sqrt(attr(.fitted, "var")),
    .fitted = as.numeric(.fitted)
  )
```

```{r}
#| code-fold: true

bind_rows(
  Original = preds_1,
  `With squared term` = preds_2,
  .id = "mdl"
) |>
  ggplot(aes(y = .resid, x = Age)) +
  geom_point(color = "goldenrod", alpha = .2) +
  facet_wrap(vars(mdl)) +
  labs(
    x = "Centered Age",
    title = glue("Residuals vs centered age before and after addition of ",
                    "squared age variable.")
  )
```

It _does_ appear better. The AIC and BIC also decreased, meaning the second model is more efficient.

```{r}
#| code-fold: true

bind_rows(
  fit_1 = glance(fit_1),
  fit_2 = glance(fit_2),
  .id = "Model Fit"
) |>
  gt::gt()
```

Add interactions between `Age` and each of the demographic factor variables, one at a time, to test whether they moderate the relationship between age and DBP. Start with race/ethnicity, and perform a Wald test of the null hypothesis that all eight parameters are simultaneously equal to zero.

```{r}
fit_3 <- svyglm(
  BPDia1 ~ Race1*Age + Race1*I(Age^2) + Gender,
  design = nhanes_des,
  subset = (Age18p == 1)
)

regTermTest(fit_3, ~ Race1:Age + Race1:I(Age^2))
```

None of the interaction estimators have a *p*-value < .05, and the Wald test fails to reject the null hypothesis. How about interacting with gender? This time the interactions are significant and the Wald test rejects the null hypothesis.

```{r}
fit_4 <- svyglm(
  BPDia1 ~ Race1 + Gender*Age + Gender*I(Age^2),
  design = nhanes_des,
  subset = (Age18p == 1)
)

summary(fit_4, df.resid = degf(nhanes_des))
regTermTest(fit_4, ~ Gender:Age + Gender:I(Age^2))
```

Plot marginal predicted values from DBP to assess variability in the relationship between age and DBP depending on race/ethnicity and gender. African Americans have a higher acceleration in DBP until middle age, then the gap decreases. 

```{r}
#| code-fold: true
bind_rows(
  `Race/Ethnicity` = fit_4 |>
    predict_response(terms = c("Age [-30:30 by=5]", "Race1")) |>
    as_tibble(),
  Gender = fit_4 |>
    predict_response(terms = c("Age [-30:30 by=5]", "Gender")) |>
    as_tibble(),
  .id = "Covar"
) |>
  mutate(group = fct_relevel(group, c("female", "male"), after = 0)) |>
  ggplot(aes(x = x, y = predicted, color = group)) + 
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.8) +
  geom_point() +
  geom_line() +
  labs(
    x = "Centered Age", y = "Marginal predicted DBP", color = NULL,
    title = "Marginal predicted values from regression model.",
    subtitle = "Interaction of age and race/ethnicity"
  ) +
  facet_wrap(vars(Covar)) +
  theme(legend.position = "top", legend.justification = "left")
```

### Model Diagnostics

The residuals vs fitted plot varies equally around 0 in a straight horizontal line (linearity and homoscedasticity). The Q-Q residuals plot is bowed at the high end, so prediction intervals maybe be inaccurate. The equal spread along the horizontal of the scale-location plot supports homoscedasticity. The residuals vs leverage plot identifies a few influential observations.

```{r}
par(mfrow = c(2, 2))
plot(fit_4, labels.id = NULL)
```

Check for influential points and outliers.

```{r}
library(MASS)
cooks_d4 <- svyCooksD(fit_4, doplot = TRUE)
```
